feat: extract CLI runtime kernel (0.64.0)
